refactor(Masonry): Update brick view to pull raw props
