+++ modules/masonry/src/brick/view/components/expression.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import type { TBrickRenderPropsExpression } from '../../@types/brick';
+import type { TBrickRenderPropsExpression } from '../../../@types/brick';
 import { BrickWrapper } from './BrickWrapper';
 import type { TConnectionPoints } from '../utils/common';
